feat: MyPresentation 제거
- MainPage에서 프로젝트 목록을 띄움
- SideBar 수정
- Router 수정
