+++ client/src/utils/api/user.js
@@ -1,23 +1,20 @@
 import { getIdToken } from '../auth/auth';
 import { UserModel } from '../../models/userModel';
 
-/// Nest에서 uid로 유저 정보를 가져옴.
+/// 유저 조회
 export async function fetchUserByUid(uid) {
     const token = await getIdToken();
     const res = await fetch(`/api/user/${uid}`, {
         headers: { Authorization: `Bearer ${token}` },
     });
-
     if (!res.ok) throw new Error('사용자 정보 요청 실패');
-
     const json = await res.json();
-    return UserModel(uid, json);
+    return UserModel(json.uid, json);
 }
 
-/// Nest에서 user의 [updates]를 업데이트함.
+/// 유저 정보 업데이트
 export async function updateUser(updates) {
     const token = await getIdToken();
-
     const res = await fetch('/api/user', {
         method: 'PATCH',
         headers: {
@@ -26,7 +23,28 @@ export async function updateUser(updates) {
         },
         body: JSON.stringify(updates),
     });
-
     if (!res.ok) throw new Error('사용자 정보 업데이트 실패');
+    return await res.json();
+}
+
+/// 유저 초기화 (서버에서 누락 필드 채우기 포함)
+export async function initUser() {
+    const token = await getIdToken();
+    const res = await fetch('/api/user/init', {
+        method: 'POST',
+        headers: { Authorization: `Bearer ${token}` },
+    });
+    if (!res.ok) throw new Error('유저 초기화 실패');
+    return await res.json();
+}
+
+/// 유저 삭제
+export async function deleteUser() {
+    const token = await getIdToken();
+    const res = await fetch('/api/user', {
+        method: 'DELETE',
+        headers: { Authorization: `Bearer ${token}` },
+    });
+    if (!res.ok) throw new Error('유저 삭제 실패');
     return await res.json();
 }
